// src/models/player-model-interface.ts
export interface IPlayerModel {
    id: number;
    name: string;
    club: string;
    nationality: string;
    position: string;
    statistiscs: {
        Overall: number;
        Pace: number;
        Shooting: number;
        Passing: number;
        Dribbling: number;
        Defending: number;
        Physical: number;
    }
}